requirements.txt
# Core AI orchestration
langgraph>=1.1.6
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Vector store
chromadb>=0.6.0

# Evaluation
ragas>=0.4.3
openai>=1.0.0

# Ollama Python SDK (for verify_models.py and direct API calls)
ollama>=0.4.0

# Demo UI
streamlit>=1.44.0

# Utilities
python-dotenv>=1.0.1
pandas>=2.2.0
pyyaml>=6.0.1
